wip: add membership support for leaving nodes
