+++ sn/src/node/core/messaging/handling/agreement.rs
@@ -146,76 +146,13 @@ impl Node {
         node_state: NodeState,
         sig: KeyedSig,
     ) -> Result<Vec<Cmd>> {
-        let mut cmds = vec![];
-        // return Ok(cmds);
-        let signature = sig.signature.clone();
-
-        // cmds.extend(self.propose_membership_change(node_state.to_msg()).await?);
-        let signed_node_state = SectionAuth {
-            value: node_state.clone(),
-            sig,
-        };
-
-        if !self
-            .network_knowledge
-            .update_member(signed_node_state.clone())
-            .await
-        {
-            info!(
-                "{}: {} at {}",
-                LogMarker::IgnoredNodeAsOffline,
-                node_state.name(),
-                node_state.addr()
-            );
-            return Ok(cmds);
-        }
-
         info!(
-            "{}: {} at {}",
-            LogMarker::AcceptedNodeAsOffline,
+            "Agreement - proposing membership change with node offline: {} at {}",
             node_state.name(),
             node_state.addr()
         );
 
-        // If this is an Offline agreement where the new node state is Relocated,
-        // we then need to send the Relocate msg to the peer attaching the signed NodeState
-        // containing the relocation details.
-        if node_state.is_relocated() {
-            cmds.push(
-                self.send_relocate(*node_state.peer(), signed_node_state)
-                    .await?,
-            );
-        }
-
-        let churn_id = ChurnId(signature.to_bytes().to_vec());
-        cmds.extend(self.relocate_peers(churn_id, BTreeSet::default()).await?);
-
-        let result = self.promote_and_demote_elders().await?;
-        if result.is_empty() {
-            // Send AE-Update to Adults of our section
-            let our_adults = self.network_knowledge.adults().await;
-            let our_prefix = self.network_knowledge.prefix().await;
-            let our_section_pk = self.network_knowledge.section_key().await;
-            cmds.extend(
-                self.send_ae_update_to_nodes(our_adults, &our_prefix, our_section_pk)
-                    .await,
-            );
-        }
-
-        cmds.extend(result);
-
-        self.liveness_retain_only(
-            self.network_knowledge
-                .adults()
-                .await
-                .iter()
-                .map(|peer| peer.name())
-                .collect(),
-        )
-        .await?;
-        *self.joins_allowed.write().await = true;
-
-        Ok(cmds)
+        self.propose_membership_change(node_state.to_msg()).await
     }
 
     #[instrument(skip(self), level = "trace")]
