+++ testnet/bin.rs
@@ -47,7 +47,7 @@ const SAFE_NODE_EXECUTABLE: &str = "sn_node.exe";
 
 const BASE_TRACING_DIRECTIVES: &str = "testnet=info,sn_launch_tool=debug";
 const NODES_DIR: &str = "local-test-network";
-const DEFAULT_INTERVAL: &str = "500";
+const DEFAULT_INTERVAL: &str = "1000";
 const DEFAULT_NODE_COUNT: u32 = 30;
 
 #[derive(Debug, StructOpt)]
